# common/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Mocker service API in Go using Gin framework.",
        "title": "Golang Template API",
        "contact": {}
    },
    "host": "localhost:8000",
    "basePath": "/mocker",
    "paths": {
...
        "/playerInfoSync": {
            "post": {
                "description": "Sync's all info about player from the DB.",
                "summary": "Sync Player Info.",
                "parameters": [
                    {
                        "description": "Player Info Sync Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.PlayerInfoSyncRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.PlayerData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/playerUpdateBalance": {
            "post": {
                "description": "Update balance for given player.",
                "summary": "Update Player Balance.",
                "parameters": [
                    {
                        "description": "Player Update Balance",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.PlayerUpdateBalanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.PlayerUpdateBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "config.PlayerData": {
            "type": "object",
            "properties": {
                "balances": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "balance": {
                                "type": "integer"
                            },
                            "publisherProductId": {
                                "type": "string"
                            }
                        }
                    }
                },
                "focus": {
                    "type": "object",
                    "properties": {
                        "publisherBundleId": {
                            "type": "string"
                        }
                    }
                },
                "offers": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "offerPlayerAvailability": {
                                "type": "integer"
                            },
                            "productsSequence": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "index": {
                                            "type": "integer"
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "badgeText": {
                                                        "type": "string"
                                                    },
                                                    "publisherProductId": {
                                                        "type": "string"
                                                    },
                                                    "quantity": {
                                                        "type": "integer"
                                                    },
                                                    "rarityProductInfo": {
                                                        "type": "object",
                                                        "properties": {
                                                            "rarityText": {
                                                                "type": "string"
                                                            },
                                                            "stars": {
                                                                "type": "integer"
                                                            },
                                                            "tooltipText": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            },
                            "publisherOfferId": {
                                "type": "string"
                            }
                        }
                    }
                },
                "playerName": {
                    "type": "string"
                },
                "playerProfileImage": {
                    "type": "string"
                },
                "publisherPlayerId": {
                    "type": "string"
                },
                "segments": {
                    "type": "array",
                    "items": {}
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "config.PlayerInfoSyncRequest": {
            "type": "object",
            "properties": {
                "playerId": {
                    "type": "string"
                }
            }
        },
        "config.PlayerUpdateBalanceRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "actionStatus": {
                    "type": "string"
                },
                "appChargePaymentId": {
                    "type": "string"
                },
                "bundleId": {
                    "type": "string"
                },
                "bundleName": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "gameId": {
                    "type": "string"
                },
                "originalPriceInDollar": {
                    "type": "integer"
                },
                "playerId": {
                    "type": "string"
                },
                "priceInCents": {
                    "type": "integer"
                },
                "priceInDollar": {
                    "type": "number"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.Product"
                    }
                },
                "purchaseDateAndTimeUtc": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "subTotal": {
                    "type": "number"
                },
                "tax": {
                    "type": "number"
                }
            }
        },
        "config.PlayerUpdateBalanceResponse": {
            "type": "object",
            "properties": {
                "publisherPurchaseId": {
                    "type": "string"
                },
                "purchaseTime": {
                    "type": "string"
                }
            }
        },
        "config.Product": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                }
            }
        }
    }
}